fixes windows crash on startup

+++ src/Core/Context.cpp
@@ -137,7 +137,7 @@ Context::Context(MainWindow *mainWindow): mainWindow(mainWindow)
     isCompareIntervals = isCompareDateRanges = false;
     isRunning = isPaused = false;
 
-    connect(this, SIGNAL(loadProgress(QString, double)), mainWindow, SLOT(loadProgress(QString, double)));
+    connect(this, SIGNAL(loadProgress(QString, int)), mainWindow, SLOT(loadProgress(QString, int)));
 
 #ifdef GC_HAS_CLOUD_DB
     cdbChartListDialog = NULL;
